+++ packages/ai-terminal/src/browser/ai-terminal-frontend-module.ts
@@ -1,0 +1,34 @@
+// *****************************************************************************
+// Copyright (C) 2024 EclipseSource GmbH.
+//
+// This program and the accompanying materials are made available under the
+// terms of the Eclipse Public License v. 2.0 which is available at
+// http://www.eclipse.org/legal/epl-2.0.
+//
+// This Source Code may also be made available under the following Secondary
+// Licenses when the conditions for such availability set forth in the Eclipse
+// Public License v. 2.0 are satisfied: GNU General Public License, version 2
+// with the GNU Classpath Exception which is available at
+// https://www.gnu.org/software/classpath/license.html.
+//
+// SPDX-License-Identifier: EPL-2.0 OR GPL-2.0-only WITH Classpath-exception-2.0
+// *****************************************************************************
+
+import { Agent } from '@theia/ai-core/lib/common';
+import { CommandContribution, MenuContribution } from '@theia/core';
+import { KeybindingContribution } from '@theia/core/lib/browser';
+import { ContainerModule } from '@theia/core/shared/inversify';
+import { AiTerminalAgent } from './ai-terminal-agent';
+import { AiTerminalCommandContribution } from './ai-terminal-contribution';
+
+import '../../src/browser/style/ai-terminal.css';
+
+export default new ContainerModule(bind => {
+    bind(AiTerminalCommandContribution).toSelf().inSingletonScope();
+    for (const identifier of [CommandContribution, MenuContribution, KeybindingContribution]) {
+        bind(identifier).toService(AiTerminalCommandContribution);
+    }
+
+    bind(AiTerminalAgent).toSelf().inSingletonScope();
+    bind(Agent).toService(AiTerminalAgent);
+});